backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
yt-dlp>=2023.3.4
youtube-transcript-api>=0.5.0
translate>=3.6.1
python-multipart>=0.0.5
pydantic>=1.8.2
httpx>=0.24.1
whisper-openai>=20231117
requests>=2.31.0
googletrans==3.1.0a0 